refactor the bubble menu
